+++ Changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [] - 2022-1-10
 ### Mejorado
 - Mensajes de ayuda de comandos
+- Mensaje de ayuda cuando se usa el nombre de un comando como el nombre de un sticker
 
 ## [3.3] - 2022-1-10
 ### Añadido